move content types into a collected folder and remove the unused withDocument
