+++ src/renderer/components/content-types/PdfContent.css
