+++ src/renderer/components/content-types/PdfContent.tsx
