+++ src/renderer/components/content-types/TextContent.tsx
