+++ src/renderer/components/content-types/ThreadContent.tsx
