+++ src/renderer/components/content-types/board/Board.css
