+++ src/renderer/components/content-types/board/BoardCard.css
