+++ src/renderer/components/content-types/board/BoardGrid.ts
