+++ src/renderer/components/content-types/board/BoardInList.tsx
