+++ src/renderer/components/content-types/board/index.ts
