+++ src/renderer/components/content-types/contact/Avatar.css
