+++ src/renderer/components/content-types/contact/ContactEditor.css
