+++ src/renderer/components/content-types/contact/ContactInVarious.css
