+++ src/renderer/components/content-types/workspace/Omnibox.css
