+++ src/renderer/components/content-types/workspace/Share.tsx
