+++ src/renderer/components/content-types/workspace/TitleBar.css
