+++ src/renderer/components/content-types/workspace/TitleBar.tsx
