+++ src/renderer/components/content-types/workspace/Workspace.tsx
